Upgrade Laravel 5.7 to 5.8

+++ database/migrations/2019_10_02_170456_create_blog_categories_table.php
@@ -14,8 +14,8 @@ class CreateBlogCategoriesTable extends Migration
     public function up()
     {
         Schema::create('blog_categories', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('parent_id')->unsigned()->default(1)->comment('Категория может быть вложенной');
+            $table->bigIncrements('id');
+            $table->bigInteger('parent_id')->unsigned()->default(1)->comment('Категория может быть вложенной');
 
             $table->string('slug')->unique()->comment('уникальный title в транслите для url');
             $table->string('title');
